+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Show error message when gateway callback returns an error response.
+
 ## [2.9.0] - 2021-05-10
 ### Changed
 - Capture first route param after `/checkout` endpoint.